fixed number of digit on the commands

+++ Ocem/OcemE642X.h
@@ -302,7 +302,7 @@ namespace common{
              @param value value
              @return 0 if success 
              */
-            int setThreashold(int channel,int value,uint32_t timeo_ms);
+            int setThreashold(int channel,float value,uint32_t timeo_ms);
